+++ lib/flipper.rb
@@ -194,5 +194,6 @@ require 'flipper/types/percentage'
 require 'flipper/types/percentage_of_actors'
 require 'flipper/types/percentage_of_time'
 require 'flipper/typecast'
+require 'flipper/version'
 
 require "flipper/railtie" if defined?(Rails::Railtie)